+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/compose/state/SelectLocationUiState.kt
@@ -1,0 +1,10 @@
+package net.mullvad.mullvadvpn.compose.state
+
+import net.mullvad.mullvadvpn.relaylist.RelayCountry
+import net.mullvad.mullvadvpn.relaylist.RelayItem
+
+sealed interface SelectLocationUiState {
+    object Loading : SelectLocationUiState
+    data class ShowData(val countries: List<RelayCountry>, val selectedRelay: RelayItem?) :
+        SelectLocationUiState
+}